\name{OAPB-package}
\alias{OAPB-package}
\alias{OAPB}
\docType{package}
\title{
\packageTitle{OAPB}
}
\description{
\packageDescription{OAPB}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{OAPB}
\packageIndices{OAPB}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{OAPB}

Maintainer: \packageMaintainer{OAPB}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
 result=RunSimulation(list(PSOD.Finite.qnorm,partial(PSOD.Finite.qnorm,q=Inf),PSOD.RKHS))
}
